protect CUD operation for products

+++ routes/product.routes.js
@@ -57,6 +57,7 @@ router.post("/products/upload-image", isAuthenticated, fileUploader.single("imag
 // POST /api/products/create  -  Creates a new product
 router.post("/products/create", isAuthenticated, (req, res, next) => {
     const { title, description, price, image_URL, specs, rating } = req.body;
+    const isAdmin = req.payload.admin
 
     const newProduct = {
       title,
@@ -67,62 +68,68 @@ router.post("/products/create", isAuthenticated, (req, res, next) => {
       rating,
     };
 
-    Product.create(newProduct)
-      .then((response) => res.json(response))
-      .catch((err) => {
-        console.log("error creating a new product...", err);
-        res.status(500).json({
-          message: "error creating a new product",
-          error: err,
+    if(isAdmin){
+      Product.create(newProduct)
+        .then((response) => res.json(response))
+        .catch((err) => {
+          console.log("error creating a new product...", err);
+          res.status(500).json({
+            message: "error creating a new product",
+            error: err,
+          });
         });
-      });
+    }
   }
 );
 
 // PUT  /api/products/:productId  -  Updates a specific product by id
 router.put("/products/:productId", isAuthenticated, (req, res, next) => {
-  //
   const { productId } = req.params;
+  const isAdmin = req.payload.admin
 
   if (!mongoose.Types.ObjectId.isValid(productId)) {
     res.status(400).json({ message: "Specified id is not valid" });
     return;
   }
-
-  Product.findByIdAndUpdate(productId, req.body, { new: true })
-    .then((updatedProduct) => res.json(updatedProduct))
-    .catch((err) => {
-      console.log("error updating specific product...", err);
-      res.status(500).json({
-        message: "error updating specific product",
-        error: err,
+  if (isAdmin){
+    Product.findByIdAndUpdate(productId, req.body, { new: true })
+      .then((updatedProduct) => res.json(updatedProduct))
+      .catch((err) => {
+        console.log("error updating specific product...", err);
+        res.status(500).json({
+          message: "error updating specific product",
+          error: err,
+        });
       });
-    });
+  }
 });
 
 // DELETE  /api/products/:productId  -  Deletes a specific product by id
 router.delete("/products/:productId", isAuthenticated, (req, res, next) => {
-  ///
   const { productId } = req.params;
+  const isAdmin = req.payload.admin
 
   if (!mongoose.Types.ObjectId.isValid(productId)) {
     res.status(400).json({ message: "Specified id is not valid" });
     return;
   }
 
-  Product.findByIdAndRemove(productId)
-    .then(() =>
-      res.json({
-        message: `Product with ${productId} is removed successfully.`,
-      })
-    )
-    .catch((err) => {
-      console.log("error deleting specific product...", err);
-      res.status(500).json({
-        message: "error deleting specific product",
-        error: err,
+  if (isAdmin){
+    Product.findByIdAndRemove(productId)
+      .then(() =>
+        res.json({
+          message: `Product with ${productId} is removed successfully.`,
+        })
+      )
+      .catch((err) => {
+        console.log("error deleting specific product...", err);
+        res.status(500).json({
+          message: "error deleting specific product",
+          error: err,
+        });
       });
-    });
+  }
+
 });
 
 module.exports = router;
